Add files via upload
This project in a group of 4 we created a library database using the program Eclipse. Created a login setup in which once the user logs in has multiple options. The user can add books, search books and delete books from database. There is also an admin menu where the user can delete, add or modify existing users.