Add files via upload
In a group of 4 I used the program Eclipse to create a library database project. In this project I created a login with username and password. Once the user logs in the user has access to either add, search or delete a book from database. The user can also access the admin menu which gives them the access to the add, delete or modify an existing user.